INFUND-7367 - Refactors ApplicationStatistics/CompetitionInviteStatistics to use ints rather than longs
Former-commit-id: 85b9224c32d25a66c6266d24618850907290d8f9

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/assessment/transactional/AssessmentService.java
@@ -34,7 +34,7 @@ public interface AssessmentService {
     @SecuredBySpring(
             value = "COUNT_BY_STATE_AND_COMPETITION",
             description = "Comp admins and execs can see a count of assessments in a particular state per competition")
-    ServiceResult<Long> countByStateAndCompetition(AssessmentStates state, long competitionId);
+    ServiceResult<Integer> countByStateAndCompetition(AssessmentStates state, long competitionId);
 
     @PreAuthorize("hasPermission(#assessmentId, 'org.innovateuk.ifs.assessment.resource.AssessmentResource', 'READ_SCORE')")
     ServiceResult<AssessmentTotalScoreResource> getTotalScore(long assessmentId);
